pyproject.toml
[project]
name = "mentorme"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "django-debug-toolbar (>=6.0.0,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "crispy-bootstrap4 (>=2025.6,<2026.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "django-filter (>=25.2,<26.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "bandit (>=1.8.6,<2.0.0)",
    "isort (>=6.1.0,<7.0.0)",
    "pylint-django (>=2.6.1,<3.0.0)",
    "ruff (>=0.13.3,<0.14.0)",
    "safety (>=3.6.2,<4.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "coverage (>=7.10.7,<8.0.0)"
]
